Extract light tool camera modals
