+++ tests/test-correctness-phase2.js
@@ -125,6 +125,8 @@ const pwaAppShellAssets = [
   '/js/mobile-dashboard.js',
   '/js/light-devices.js',
   '/js/light-env.js',
+  '/js/light-tool-camera.js',
+  '/js/light-tool-camera-modals.js',
   '/js/light-tools.js',
   '/js/sun.js',
   '/js/sun-spectrum.js',
